Flatten a Crunch Cube

Crunch Cubes can be expressed as a long data frame instead of a multidimensional array. In this form each dimension of the cube is a variable and the cube values are expressed as columns for each measure. This is useful both to better understand what each entry of a cube represents, and to work with the cube result using tidyverse tools.

as_cr_tibble(x, ...)

Arguments

x
a CrunchCube

...
further arguments passed on to tibble::as_tibble()

Details

The cr_tibble class is a subclass of tibble that has extra metadata to allow ggplot::autoplot() to work. If you find that this extra metadata is getting in the way, you can use as_tibble() to get a true tibble.