Autoplot methods for Crunch Objects

The Crunch autoplot methods generate ggplots that are tailored to various Crunch objects. This allows you to visualize the object without bringing it into memory. You can select between three families of plots, which will attempt to accomodate the dimensionality of the plotted object. These plots can be further extended and customized with other ggplot methods.

# S3 method for DatetimeVariable
autoplot(x, ...)

# S3 method for NumericVariable
autoplot(x, ...)

# S3 method for CategoricalVariable
autoplot(x, ...)

# S3 method for CategoricalArrayVariable
autoplot(x, ...)

# S3 method for MultipleResponseVariable
autoplot(x, ...)

# S3 method for CrunchCube
autoplot(x, ...)

# S3 method for CrunchCubeCalculation
autoplot(x, plot_type = "dot", ...)

# S3 method for tbl_crunch_cube
autoplot(x, plot_type = c("dot", "tile", "bar"), measure)

Arguments

x
A Crunch variable or cube aggregation

...
additional plotting arguments

plot_type
One of "dot", "tile", or "bar" which indicates the plot family you would like to use. Higher dimensional plots add color coding or facets depending on the dimensionality of the data.

measure
The measure you wish to plot. This will usually be "count", the default but can also be ".unweighted_counts" or any other measure stored in the cube. If omitted, autoplot will select the first measure appearing in the data.

Value

A ggplot object.